+++ Assets/Prefabs/BaseFood/egg_food.prefab
@@ -11,10 +11,9 @@ GameObject:
   - component: {fileID: 5998635098908830834}
   - component: {fileID: 5128741936716209196}
   - component: {fileID: 6575779008971730262}
-  - component: {fileID: -6942545737338809984}
-  - component: {fileID: 8888346746106426489}
-  - component: {fileID: -7287063832135636201}
+  - component: {fileID: 4018776367589799187}
   - component: {fileID: -2591238644688280200}
+  - component: {fileID: -6942545737338809984}
   m_Layer: 0
   m_Name: egg_food
   m_TagString: food item
@@ -101,85 +100,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f8c41fee416e00542914f09b0415e4f1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  hoveredSprite: {fileID: 21300000, guid: f34bab22c27546c4fbebaa69fff4b6e1, type: 3}
-  initalMouse: {x: 0, y: 0, z: 0}
-  canDragandDrop: 1
-  isPlate: 0
   dragging: 1
-  firstCollidingObject: {fileID: 0}
---- !u!50 &-6942545737338809984
-Rigidbody2D:
-  serializedVersion: 4
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 580618499572302972}
-  m_BodyType: 0
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 1
-  m_Material: {fileID: 6200000, guid: 7d8d3e7a6b8f8c949a1c7de747671078, type: 2}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
---- !u!61 &8888346746106426489
-BoxCollider2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 580618499572302972}
-  m_Enabled: 0
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_LayerOverridePriority: 0
-  m_ForceSendLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_ForceReceiveLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_ContactCaptureLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_CallbackLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_IsTrigger: 0
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 0.89, y: 0.95}
-    newSize: {x: 0.89, y: 0.95}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 0.92485714, y: 1.0913296}
-  m_EdgeRadius: 0
---- !u!114 &-7287063832135636201
+  rigidbod: {fileID: 0}
+--- !u!114 &4018776367589799187
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -188,13 +111,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 580618499572302972}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 02d23a041d9ebc5419fee6d2e4ae2009, type: 3}
+  m_Script: {fileID: 11500000, guid: d7c5fee204b2d6d43b1b7349b0e4cd93, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  crushed: {fileID: 8297354940826436125, guid: 340748ec438b1984cbd0cfc196649fc9, type: 3}
-  cut: {fileID: 0}
-  cooked: {fileID: 0}
-  foodColor: {r: 0.6666667, g: 0.87843144, b: 0.50980395, a: 1}
+  currentColl: {fileID: 0}
+  stored: 0
 --- !u!60 &-2591238644688280200
 PolygonCollider2D:
   m_ObjectHideFlags: 0
@@ -257,3 +178,30 @@ PolygonCollider2D:
       - {x: 0.39499998, y: -0.295}
       - {x: 0.435, y: -0.205}
   m_UseDelaunayMesh: 0
+--- !u!50 &-6942545737338809984
+Rigidbody2D:
+  serializedVersion: 4
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 580618499572302972}
+  m_BodyType: 0
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 1
+  m_Material: {fileID: 6200000, guid: 7d8d3e7a6b8f8c949a1c7de747671078, type: 2}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
