+++ Assets/Prefabs/BaseFood/sheet_food.prefab
@@ -11,9 +11,9 @@ GameObject:
   - component: {fileID: 7328869825394656848}
   - component: {fileID: 3525741330097541702}
   - component: {fileID: 7330042575336150700}
-  - component: {fileID: -8211959908704455596}
+  - component: {fileID: -4431532476199755897}
   - component: {fileID: 6879090268991256250}
-  - component: {fileID: 5778548249425939243}
+  - component: {fileID: -8211959908704455596}
   m_Layer: 0
   m_Name: sheet_food
   m_TagString: food item
@@ -100,37 +100,22 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f8c41fee416e00542914f09b0415e4f1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  hoveredSprite: {fileID: 21300000, guid: 8099f9ed02dac9b40b7317dc71ec927f, type: 3}
-  initalMouse: {x: 0, y: 0, z: 0}
   dragging: 1
-  firstCollidingObject: {fileID: 0}
---- !u!50 &-8211959908704455596
-Rigidbody2D:
-  serializedVersion: 4
+  rigidbod: {fileID: 0}
+--- !u!114 &-4431532476199755897
+MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6026397797376136479}
-  m_BodyType: 0
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 1
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d7c5fee204b2d6d43b1b7349b0e4cd93, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  currentColl: {fileID: 0}
+  stored: 0
 --- !u!60 &6879090268991256250
 PolygonCollider2D:
   m_ObjectHideFlags: 0
@@ -218,19 +203,30 @@ PolygonCollider2D:
       - {x: 0.825, y: -1.225}
       - {x: 0.825, y: -0.865}
   m_UseDelaunayMesh: 0
---- !u!114 &5778548249425939243
-MonoBehaviour:
+--- !u!50 &-8211959908704455596
+Rigidbody2D:
+  serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6026397797376136479}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 02d23a041d9ebc5419fee6d2e4ae2009, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  crushed: {fileID: 8297354940826436125, guid: 8af8e357dc883664ab5909f7ff7a6245, type: 3}
-  cut: {fileID: 0}
-  cooked: {fileID: 0}
-  foodColor: {r: 0.23137257, g: 0.31764707, b: 0.67058825, a: 1}
+  m_BodyType: 0
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 1
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
