+++ Assets/Prefabs/BaseFood/tentacle_food.prefab
@@ -11,10 +11,9 @@ GameObject:
   - component: {fileID: 4245082900709541092}
   - component: {fileID: 6765793632034582510}
   - component: {fileID: 4130368508489723484}
-  - component: {fileID: 5756428243654078875}
-  - component: {fileID: -6235215649513243321}
+  - component: {fileID: -7879753483142807167}
   - component: {fileID: 3323012342824336015}
-  - component: {fileID: 6825458188127815068}
+  - component: {fileID: 7004981561572283162}
   m_Layer: 0
   m_Name: tentacle_food
   m_TagString: food item
@@ -29,13 +28,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4281459659418726516}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &6765793632034582510
 SpriteRenderer:
@@ -101,82 +100,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f8c41fee416e00542914f09b0415e4f1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  hoveredSprite: {fileID: 21300000, guid: c5b9dd4d6ee9de640ab2d112482b664c, type: 3}
-  initalMouse: {x: 0, y: 0, z: 0}
   dragging: 1
-  firstCollidingObject: {fileID: 0}
---- !u!61 &5756428243654078875
-BoxCollider2D:
+--- !u!114 &-7879753483142807167
+MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4281459659418726516}
-  m_Enabled: 0
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_LayerOverridePriority: 0
-  m_ForceSendLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_ForceReceiveLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_ContactCaptureLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_CallbackLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  m_IsTrigger: 0
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 2.35, y: 1.66}
-    newSize: {x: 2.35, y: 1.66}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 2.4318469, y: 1.7013605}
-  m_EdgeRadius: 0
---- !u!50 &-6235215649513243321
-Rigidbody2D:
-  serializedVersion: 4
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4281459659418726516}
-  m_BodyType: 0
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 1
-  m_Material: {fileID: 0}
-  m_IncludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_ExcludeLayers:
-    serializedVersion: 2
-    m_Bits: 0
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d7c5fee204b2d6d43b1b7349b0e4cd93, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!60 &3323012342824336015
 PolygonCollider2D:
   m_ObjectHideFlags: 0
@@ -268,19 +204,30 @@ PolygonCollider2D:
       - {x: 1.0749999, y: -0.41}
       - {x: 1.155, y: -0.22999999}
   m_UseDelaunayMesh: 0
---- !u!114 &6825458188127815068
-MonoBehaviour:
+--- !u!50 &7004981561572283162
+Rigidbody2D:
+  serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4281459659418726516}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 02d23a041d9ebc5419fee6d2e4ae2009, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  crushed: {fileID: 8297354940826436125, guid: aba9ed37d2f0e8b4ba07129d0df2f608, type: 3}
-  cut: {fileID: 0}
-  cooked: {fileID: 0}
-  foodColor: {r: 0.5764706, g: 0.7137255, b: 0.70980394, a: 1}
+  m_BodyType: 0
+  m_Simulated: 1
+  m_UseFullKinematicContacts: 0
+  m_UseAutoMass: 0
+  m_Mass: 1
+  m_LinearDrag: 0
+  m_AngularDrag: 0.05
+  m_GravityScale: 1
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 256
+  m_Interpolate: 0
+  m_SleepingMode: 1
+  m_CollisionDetection: 0
+  m_Constraints: 0
