+++ Assets/Prefabs/new kitchen.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a2ed6dcbda0dd5d49835e025c4cc65e4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
