+++ Assets/Prefabs/tray.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &6079846946831491505
+--- !u!1 &1925612710533241458
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,41 +8,41 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 3002744647587005214}
-  - component: {fileID: 5465089782519685911}
-  - component: {fileID: 8781841345362289858}
-  - component: {fileID: 6923007331855906059}
-  - component: {fileID: 7807008616026827326}
-  - component: {fileID: 6525950345671681695}
+  - component: {fileID: 3960107473091969522}
+  - component: {fileID: 6127532923372566551}
+  - component: {fileID: 1160314042551011212}
+  - component: {fileID: 2957144155675455592}
+  - component: {fileID: 7644371535918866613}
+  - component: {fileID: 4655442506656536633}
   m_Layer: 0
-  m_Name: leaf_food
-  m_TagString: food item
+  m_Name: tray
+  m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &3002744647587005214
+--- !u!4 &3960107473091969522
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6079846946831491505}
+  m_GameObject: {fileID: 1925612710533241458}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalPosition: {x: 4.33, y: 1.66, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 0}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &5465089782519685911
+--- !u!212 &6127532923372566551
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6079846946831491505}
+  m_GameObject: {fileID: 1925612710533241458}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -74,55 +74,56 @@ SpriteRenderer:
   m_AutoUVMaxDistance: 0.5
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 992712277
-  m_SortingLayer: 6
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: 21300000, guid: 4f421b8ac490d60458395502d53273c6, type: 3}
+  m_Sprite: {fileID: 21300000, guid: 1065a221de4ce9a4cb82e44e0d0b3d03, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
-  m_Size: {x: 2.02, y: 2.05}
+  m_Size: {x: 1, y: 1}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!114 &8781841345362289858
+--- !u!114 &1160314042551011212
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6079846946831491505}
+  m_GameObject: {fileID: 1925612710533241458}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: f8c41fee416e00542914f09b0415e4f1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dragging: 1
+  dragging: 0
   rigidbod: {fileID: 0}
---- !u!114 &6923007331855906059
+--- !u!114 &2957144155675455592
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6079846946831491505}
+  m_GameObject: {fileID: 1925612710533241458}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d7c5fee204b2d6d43b1b7349b0e4cd93, type: 3}
+  m_Script: {fileID: 11500000, guid: b448c812fd3fafa49b6e3c1e5374e567, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  currentColl: {fileID: 0}
-  stored: 0
---- !u!60 &7807008616026827326
-PolygonCollider2D:
+  maxCapacity: 4
+  currentCap: 0
+  position: {x: 0, y: 0, z: 0}
+--- !u!61 &7644371535918866613
+BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6079846946831491505}
+  m_GameObject: {fileID: 1925612710533241458}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
@@ -152,64 +153,23 @@ PolygonCollider2D:
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 2.02, y: 2.05}
-    newSize: {x: 2.02, y: 2.05}
+    oldSize: {x: 3.83, y: 0.77}
+    newSize: {x: 1, y: 1}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
     adaptiveTiling: 0
   m_AutoTiling: 0
-  m_Points:
-    m_Paths:
-    - - {x: 0.22999999, y: -0.525}
-      - {x: 0.39, y: -0.475}
-      - {x: 0.59, y: -0.385}
-      - {x: 0.82, y: -0.225}
-      - {x: 0.78999996, y: -0.155}
-      - {x: 0.97999996, y: 0.055}
-      - {x: 1.01, y: 0.13499999}
-      - {x: 1.01, y: 0.265}
-      - {x: 0.95, y: 0.285}
-      - {x: 0.87, y: 0.575}
-      - {x: 0.87, y: 0.745}
-      - {x: 0.78999996, y: 0.79499996}
-      - {x: 0.72999996, y: 0.695}
-      - {x: 0.68, y: 0.685}
-      - {x: 0.48999998, y: 0.78499997}
-      - {x: 0.35999998, y: 0.965}
-      - {x: 0.31, y: 0.965}
-      - {x: 0.14999999, y: 0.755}
-      - {x: 0.14999999, y: 0.84499997}
-      - {x: 0.12, y: 0.91499996}
-      - {x: 0.03, y: 0.965}
-      - {x: -0.11, y: 1.025}
-      - {x: -0.26999998, y: 1.025}
-      - {x: -0.29, y: 0.995}
-      - {x: -0.31, y: 0.945}
-      - {x: -0.41, y: 0.745}
-      - {x: -0.55, y: 0.59499997}
-      - {x: -0.66999996, y: 0.59499997}
-      - {x: -0.89, y: 0.355}
-      - {x: -1.01, y: 0.155}
-      - {x: -1.01, y: -0.225}
-      - {x: -0.98999995, y: -0.275}
-      - {x: -0.9, y: -0.475}
-      - {x: -0.78999996, y: -0.635}
-      - {x: -0.65999997, y: -0.775}
-      - {x: -0.55, y: -0.875}
-      - {x: -0.31, y: -1.005}
-      - {x: -0.17999999, y: -1.005}
-      - {x: -0.13, y: -0.98499995}
-      - {x: -0.07, y: -0.805}
-      - {x: 0.13, y: -0.645}
-  m_UseDelaunayMesh: 0
---- !u!50 &6525950345671681695
+  serializedVersion: 2
+  m_Size: {x: 1, y: 1}
+  m_EdgeRadius: 0
+--- !u!50 &4655442506656536633
 Rigidbody2D:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6079846946831491505}
+  m_GameObject: {fileID: 1925612710533241458}
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
