+++ Assets/Prefabs/tray.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a61ea0b79449c8847a16ffb49df7e224
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
