+++ Assets/Scripts/FoodSpawn.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+public class FoodSpawn : MonoBehaviour
+{
+    [SerializeField]
+    private GameObject foodItem;
+
+    private void OnMouseDown()
+    {
+        Vector3 spawnPos = Camera.main.ScreenToWorldPoint(Input.mousePosition);
+        spawnPos.z = 0;
+
+        Instantiate(foodItem, spawnPos, transform.rotation);
+    }
+}
