+++ Assets/Scripts/FoodSpawn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: adacf0a686d1f2544a2266d24632533b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
