+++ Assets/Scripts/Ingredient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7c5fee204b2d6d43b1b7349b0e4cd93
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
