+++ Assets/Scripts/OldDragAndDrop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7b21a305b20a5564ea998d839fc3f62d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
