+++ Assets/Scripts/OldStorage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ddfa94aa851abb489d35ec2ff1ee4b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
